Move match card to a separate file
