+++ src/routes/event/index.tsx
@@ -1,25 +1,8 @@
 import { h } from 'preact'
 import LoadData from '../../load-data'
 import Page from '../../components/page'
-import { getEventMatches, MatchInfo } from '../../api'
-import { formatTeamNumber } from '../../utils'
-
-interface MatchCardProps {
-  match: MatchInfo
-}
-
-const MatchCard = ({ match }: MatchCardProps) => (
-  <div>
-    <h1>{match.key}</h1>
-    <h2>{match.time}</h2>
-    <h2 style={{ color: 'red' }}>
-      {match.redAlliance.map(formatTeamNumber).join(' ')}
-    </h2>
-    <h2 style={{ color: 'blue' }}>
-      {match.blueAlliance.map(formatTeamNumber).join(' ')}
-    </h2>
-  </div>
-)
+import { getEventMatches } from '../../api'
+import { MatchCard } from '../../components/match-card'
 
 interface EventProps {
   eventKey: string
@@ -33,7 +16,7 @@ const Event = ({ eventKey }: EventProps) => (
         console.log(matches)
         return (
           <div>
-            Hello {eventKey}{' '}
+            Hello {eventKey}
             {matches && matches.map(m => <MatchCard match={m} />)}
           </div>
         )
